+++ CHANGES.rst
@@ -2,10 +2,18 @@
  Changes
 =========
 
-2.4.3 (unreleased)
+3.0.0 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Update to zope.interface 5.3. and zope.component 5.0.
+  This lets the ``BTreeAdapterRegistry``
+  and ``BTreeLocalSiteManager`` work much more smoothly and scale
+  better.
+
+  .. important::
+
+     The automatic conversion from dicts to BTrees has been removed.
+     You must explicitly migrate by calling ``BTreeLocalSiteManager.rebuild()``.
 
 
 2.4.2 (2021-03-01)